requirements.txt
# Discord Bot Core
discord.py>=2.3.2
python-dotenv>=1.0.0

# Database
aiosqlite>=0.19.0
SQLAlchemy>=2.0.23

# Data Integration
gspread>=5.12.0
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Web Dashboard
flask>=3.0.0
flask-cors>=4.0.0
flask-login>=0.6.3

# Utilities
aiohttp>=3.9.1
requests>=2.31.0
pyyaml>=6.0.1
python-dateutil>=2.8.2

# Data Processing & Analytics
pandas>=2.1.4
matplotlib>=3.8.2

# Security
cryptography>=41.0.7

# Moderation (Optional ML - connects to model.py)
numpy>=1.26.2
scikit-learn>=1.3.2

# Scheduling
APScheduler>=3.10.4

# Logging
colorlog>=6.8.0